requirements.txt
Flask
numpy
keras
tensorflow
opencv-python-headless
Pillow